requirements.txt
anyio==4.9.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyarrow==19.0.1
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
requests==2.32.3
requests-file==2.1.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
selectolax==0.3.28
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
tldextract==5.1.3
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
webencodings==0.5.1
